added support for videos
This reworks this to allow both videos and images to be download. It also changes the time format and rearranges the name of the files generated.

There is one issue where apparently instead of a thumbnail the actual image is uploaded and a "big" file isn't created. This causes an S3 error/404 as the file doesn't exist in S3 in that format. For these I have added exceptions to print error. There's probably a better way to account for these.